add speech component
